Buttons and windows starting to take shape
Most of the tkinter look is at least roughed out--several windows need work, and there is probably need for an initial & a final 'orientation' window. Plus it would be great to add some text entry windows that will change the text of the generated greeting--where the operator can add her name, for example, and a custom text wording.