Struggling with key binding; nice windows though
Unable to figure how to properly capture user's choices for hotkeys--they will write to the terminal but not to a proper variable that would appear in a dynamic label. Additionally, using pynput.keyboard is messing up the hotkey/wait() at the bottom. Why not have users directly enter their key choices--but this would force them to type 'Ctrl' or 'Shift' etc. for several likely keys. Maybe a combo of tkinter Text entry and Checkboxes (for those special keys) makes more sense.