Greeting works; "kill switch" in place
Needs an optimal choice of dead keys, and prior to packaging, a straightforward way to launch and stop the program--also prior to that, need to shut down logging of the greeting, and is there a way to include a 'starter' config file? If not, need the program to seamlessly write config.ini. Also, final touch needs to be granular reconfiguration of: 1-Click coords; 2- Greeting script; 3-Key binds