Calculates Policy Page & already p/u clicks
Finds screen size (in a clunky way with Tkinter--opens "root" window and then destroys it) and tries to discern where to click to open Policy Page and to dismiss notice that chat has been picked up by another librarian. Q: Will it work with that precision, or will it be necessary to write a "shotgun" function to spray clicks in the vicinity of each?